requirements.txt
# Web Arayüzü
streamlit==1.32.0
streamlit-option-menu==0.3.12
hydralit_components==1.0.10

# Video İşleme
moviepy==1.0.3
ffmpeg-python==0.2.0

# Ses Tanıma ve İşleme
openai-whisper
whisper
soundfile==0.12.1
librosa==0.10.1

# Çeviri Servisleri
deepl==1.16.1
googletrans==3.1.0a0

# Ses Sentezi (TTS)
edge-tts==6.1.9

# Yüz Analizi ve Senkronizasyon
mediapipe==0.10.9
face-alignment==1.4.1

# Yardımcı Kütüphaneler
python-dotenv==1.0.0
numpy==1.24.3

# Torch (Whisper için gerekli)
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==2.2.0+cu118
torchaudio==2.2.0+cu118 